fix: make clip_start and clip_end at same type

+++ src/check_csv/types.rs
@@ -6,7 +6,7 @@ pub struct Music {
     video_type: String,
     video_id: String,
     clip_start: Option<f32>,
-    clip_end: Option<f64>,
+    clip_end: Option<f32>,
     status: Option<u16>,
     title: String,
     artist: String,
